Crate r3_support_rza1

Supporting package for running R3 on a Renesas RZ/A1x family MPU (including RZ/A1H, which is used by the GR-PEACH development board).

Modules

_changelog_

Changelog

gpio

Pin configuration

serial

Serial driver

stdout

Standard output

Macros

sprint

Macro for printing to the serial standard output

sprintln

Macro for printing to the serial standard output, with a newline.

use_os_timer

Attach the implementation of PortTimer that is based on RZ/A1 OS Timer to a given kernel trait type. This macro also implements Timer on the kernel trait type. Requires OsTimerOptions and Gic.

Traits

OsTimerOptions

The options for use_os_timer!.