Trait r3_support_rza1::OsTimerOptions

pub trait OsTimerOptions {
    const OSTM_BASE: usize;
    const STBCR_OSTM: Option<(usize, u8)>;
    const FREQUENCY: u64;
    const FREQUENCY_DENOMINATOR: u64;
    const HEADROOM: u32;
    const INTERRUPT_OSTM_PRIORITY: InterruptPriority;
    const INTERRUPT_OSTM: InterruptNum;
}

The options for use_os_timer!.

Associated Constants

const OSTM_BASE: usize

The base address of OS Timer’s memory-mapped registers.

const STBCR_OSTM: Option<(usize, u8)>

The standby control register’s memory address and bit position used to enable the clock supply to OS Timer.

Defaults to Some((0xfcfe0428, 1)) (STBCR5.MSTP51).

const FREQUENCY: u64

The numerator of the timer clock rate of the timer.

const FREQUENCY_DENOMINATOR: u64

The denominator of the timer clock rate of the timer. Defaults to 1.

const HEADROOM: u32

The maximum permissible timer interrupt latency, measured in hardware timer cycles.

Defaults to min(FREQUENCY * 60 / FREQUENCY_DENOMINATOR, 0x40000000).

const INTERRUPT_OSTM_PRIORITY: InterruptPriority

The interrupt priority of the timer interrupt line. Defaults to 0xc0.

const INTERRUPT_OSTM: InterruptNum

OS Timer’s interrupt number.

Loading content...